requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.1.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
drf-yasg==1.20.0
drf-writable-nested
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pip==19.0.3
#psycopg2==2.8.6
psycopg2-binary
pycparser==2.20
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
setuptools==40.8.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
attrs==20.3.0
importlib-metadata==3.10.0
iniconfig==1.1.1
pluggy==0.13.1
py==1.10.0
pytest==6.2.2
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.1
pytest-django==4.1.0
coverage==5.5
django-ssl-auth==2.2.0
django-filter==2.4.0
django-environ==0.4.5
